+++ compiled/jquery-mobile-angular-adapter-1.1.1-SNAPSHOT.js
@@ -108,6 +108,31 @@ factory(window.jQuery, window.angular);
         }
     };
 
+    // navbar does not contain a refresh function, so we add it here.
+    $.mobile.navbar.prototype.refresh = function() {
+        var $navbar = this.element,
+            $navbtns = $navbar.find( "a" ),
+            iconpos = $navbtns.filter( ":jqmData(icon)" ).length ?
+                this.options.iconpos : undefined;
+
+        var list = $navbar.find("ul");
+        var listEntries = list.children("li");
+        list.removeClass(function (index, css) {
+            return (css.match(/\bui-grid-\S+/g) || []).join(' ');
+        });
+        listEntries.removeClass(function (index, css) {
+            return (css.match(/\bui-block-\S+/g) || []).join(' ');
+        });
+        list.jqmEnhanceable().grid({ grid: this.options.grid });
+
+        $navbtns.buttonMarkup({
+            corners:	false,
+            shadow:		false,
+            inline:     true,
+            iconpos:	iconpos
+        });
+    };
+
 })($);
 /**
  * This will delay the angular initialization by two nested calls to jQuery.fn.ready.
@@ -696,7 +721,7 @@ factory(window.jQuery, window.angular);
             handlers:[refreshControlgroupOnChildrenChange]
         },
         navbar:{
-            handlers:[]
+            handlers:[refreshOnChildrenChange]
         },
         dialog:{
             handlers:[]