+++ Changelog.md
@@ -5,6 +5,7 @@ Changelog
 -------------
 - Bugfixes to different elements.
 - Collapsible now has bidirectonal data binding for the `data-collapsed` attribute.
+- `navbar` automatically refreshes if children `li` elements are added / removed.
 - `paged` filter changed and simplified, so that it can be used with angular filter chaining.